requirements.txt
functions-framework==3.*
requests==2.28.2
flask==2.1.0
python-telegram-bot-raw==20.1
pymongo==4.3.3
emoji==2.2.0
fake-useragent==1.1.1
beautifulsoup4==4.11.2
google-cloud-translate==2.0.1